+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## in dev
 ### Changes (non-breaking)
+- `luid-translations-list` - `deletion-callback` attribut called before a value deletion.
 ### Bug fixes
 
 ## 3.1.18 - [release](https://github.com/LuccaSA/lucca-ui/releases/tag/3.1.18)